+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.3.1] - 2020-08-05
+###
+- Adapt to importmode API changes in pytest-6.0
+
 ## [0.3] - 2020-08-04
 ###
 - Adapt to Node API changes in pytest-5.4 (Fixes #11)